Change the is_valid_move method fill the board
I changed the way the pieces can validate the moves. The piece
will no longer be aware of it's current position, because I think
it's unnecessary information for the piece.

So in order to check if the move is valid you now need to give the
origin of the chess piece.

All the created chess pieces now have a to string function. Run the
code from main.py to see the results of this.